[print] Print Faust programs in functional form.
Example:

```
$ ./compiler/faust -I libraries/ -lang wagner examples/filtering/bandFilter.dsp
```

We also incorporate a .dir-locals.el file for emacs users.

Description of the algorithm:
-----------------------------

The idea is to convert shared expressions into let-bound
ones. To this extent, we traverse Faust trees and keep a cache of
already translated expressions.

However, we must be very careful with scopes introduced by binders, as
in general it is not sound to memoize the same syntactic expression.

Thus, every time we find a deBruijn binder we must create a new
memoization table.